quant-agent/requirements.txt
functions-framework
requests
yfinance
python-dotenv
google-cloud-storage